Move FrameworkEventSource to shared partition
and remove unused code
